requirements.txt
# Editable Git install with no remote (aiorchestrator==0.1)
-e c:\users\junior amaya\cascadeprojects\aiorchestrator
asyncpg==0.28.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.6.12
filelock==3.14.0
flatbuffers==25.2.10
fsspec==2025.2.0
greenlet==3.1.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.22.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
onnxruntime==1.20.1
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
protobuf==5.29.3
pyreadline3==3.5.4
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.0
python-json-logger==3.2.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
setuptools==75.8.0
soupsieve==2.6
sqlalchemy==2.0.23
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
urllib3==2.3.0
